+++ KitePrintSDK/src/main/java/ly/kite/checkout/DefaultPaymentFragment.java
@@ -42,16 +42,37 @@ package ly.kite.checkout;
 import android.app.Activity;
 import android.content.Intent;
 import android.os.Bundle;
+import android.util.Log;
 import android.view.LayoutInflater;
 import android.view.View;
 import android.view.ViewGroup;
-import android.widget.TextView;
+import android.widget.Button;
+import android.widget.ImageButton;
 
+import com.google.android.gms.common.api.ApiException;
+import com.google.android.gms.common.api.Status;
+import com.google.android.gms.tasks.OnCompleteListener;
+import com.google.android.gms.tasks.Task;
+import com.google.android.gms.wallet.AutoResolveHelper;
+import com.google.android.gms.wallet.CardRequirements;
+import com.google.android.gms.wallet.IsReadyToPayRequest;
+import com.google.android.gms.wallet.PaymentData;
+import com.google.android.gms.wallet.PaymentDataRequest;
+import com.google.android.gms.wallet.PaymentMethodToken;
+import com.google.android.gms.wallet.PaymentMethodTokenizationParameters;
+import com.google.android.gms.wallet.PaymentsClient;
+import com.google.android.gms.wallet.TransactionInfo;
+import com.google.android.gms.wallet.Wallet;
+import com.google.android.gms.wallet.WalletConstants;
 import com.paypal.android.sdk.payments.PayPalPayment;
 import com.paypal.android.sdk.payments.PaymentConfirmation;
 import com.paypal.android.sdk.payments.ProofOfPayment;
 import com.paypal.android.sdk.payments.ShippingAddress;
 
+import org.json.JSONException;
+import org.json.JSONObject;
+
+import java.util.Arrays;
 import ly.kite.KiteSDK;
 import ly.kite.address.Address;
 import ly.kite.catalogue.MultipleCurrencyAmounts;
@@ -75,6 +96,7 @@ public class DefaultPaymentFragment extends APaymentFragment
   static private final String  LOG_TAG                          = "DefaultPaymentFragment";
 
   static private final int     ACTIVITY_REQUEST_CODE_PAYPAL     = 23;
+  static private final int ACTIVITY_REQUEST_CODE_GOOGLE_PAY     = 27;
 
   //static private final String CARD_IO_TOKEN = "f1d07b66ad21407daf153c0ac66c09d7";
 
@@ -87,15 +109,14 @@ public class DefaultPaymentFragment extends APaymentFragment
 
   ////////// Member Variable(s) //////////
 
+  private boolean           mGooglePayAvailable;
   private boolean           mPayPalAvailable;
 
-  private TextView          mPayPalTextView;
-  private View              mPayPalView;
+  private ImageButton       mGooglePayButton;
+  private ImageButton       mPayPalButton;
+  private Button            mCreditCardButton;
 
-  private TextView          mChangeableCreditCardTextView;
-  private TextView          mFixedCreditCardTextView;
-  private TextView          mCreditCardTextView;
-
+  private PaymentsClient    mPaymentsClient;
   private ICreditCardAgent  mCreditCardAgent;
 
 
@@ -143,72 +164,34 @@ public class DefaultPaymentFragment extends APaymentFragment
     {
     View view = layoutInflater.inflate( R.layout.fragment_default_payment, container, false );
 
-
-    mPayPalTextView = (TextView)view.findViewById( R.id.paypal_text_view );
-
-    if ( mPayPalTextView == null ) mPayPalTextView = (TextView)view.findViewById( R.id.cta_bar_left_text_view );
-
-    mPayPalView = view.findViewById( R.id.paypal_view );
+    mGooglePayButton =  view.findViewById( R.id.google_pay_button );
+    mPayPalButton = view.findViewById( R.id.paypal_button );
+    mCreditCardButton = view.findViewById( R.id.credit_card_button );
 
 
-    mChangeableCreditCardTextView = (TextView)view.findViewById( R.id.credit_card_text_view );
+    // Determine what payment options are available / enabled
 
-    if ( mChangeableCreditCardTextView == null ) mChangeableCreditCardTextView = (TextView)view.findViewById( R.id.cta_bar_right_text_view );
-
-    mFixedCreditCardTextView = (TextView)view.findViewById( R.id.fixed_credit_card_text_view );
-
-
-    if ( mChangeableCreditCardTextView != null ) mCreditCardTextView = mChangeableCreditCardTextView;
-    else                                         mCreditCardTextView = mFixedCreditCardTextView;
-
-
-    // Determine if PayPal payments are available
+    mGooglePayAvailable = KiteSDK.getInstance( getActivity() ).getGooglePayPaymentsEnabled();
     mPayPalAvailable = KiteSDK.getInstance( getActivity() ).getPayPalPaymentsAvailable();
 
 
     // Set up the buttons
 
-    if ( mPayPalAvailable )
-      {
-      if ( mPayPalTextView != null )
-        {
-        mPayPalTextView.setText( R.string.kitesdk_payment_paypal_button_text);
-        mPayPalTextView.setTextColor( getResources().getColor( R.color.payment_paypal_button_text ) );
+    if ( mGooglePayAvailable ){
+      startGooglePayClient();
+    } else {
+      mGooglePayButton.setVisibility( View.GONE );
+    }
 
-        mPayPalTextView.setOnClickListener( this );
-        }
+    if ( mPayPalAvailable ){
+      mPayPalButton.setOnClickListener( this );
+    } else {
+      mPayPalButton.setVisibility( View.GONE );
+    }
 
-      if ( mPayPalView != null )
-        {
-        mPayPalView.setOnClickListener( this );
-        }
-      }
-    else
-      {
-      if ( mPayPalTextView != null )
-        {
-        mPayPalTextView.setVisibility( View.GONE );
-        }
-
-      if ( mPayPalView != null )
-        {
-        mPayPalView.setVisibility( View.GONE );
-        }
-      }
-
-
-    if ( mChangeableCreditCardTextView != null )
-      {
-      mChangeableCreditCardTextView.setTextColor( getResources().getColor( R.color.payment_credit_card_button_text ) );
-      }
-
-    mCreditCardTextView.setText( R.string.kitesdk_payment_credit_card_button_text);
-
-    mCreditCardTextView.setOnClickListener( this );
-
+    mCreditCardButton.setOnClickListener( this );
 
     getPaymentActivity().onPaymentFragmentReady();
-
 
     return ( view );
     }
@@ -222,10 +205,10 @@ public class DefaultPaymentFragment extends APaymentFragment
   @Override
   public void onEnableButtons( boolean enabled )
     {
-    if ( mPayPalTextView != null ) mPayPalTextView.setEnabled( enabled && mPayPalAvailable );
-    if ( mPayPalView     != null ) mPayPalView.setEnabled( enabled && mPayPalAvailable );
+    if ( mGooglePayButton != null ) mGooglePayButton.setEnabled( enabled && mGooglePayAvailable );
+    if ( mPayPalButton    != null ) mPayPalButton.setEnabled( enabled && mPayPalAvailable );
 
-    mCreditCardTextView.setEnabled( enabled );
+    mCreditCardButton.setEnabled( enabled );
     }
 
 
@@ -239,11 +222,11 @@ public class DefaultPaymentFragment extends APaymentFragment
     {
     if ( free )
       {
-      if ( mPayPalTextView != null ) mPayPalTextView.setVisibility( View.GONE );
-      if ( mPayPalView     != null ) mPayPalView.setVisibility( View.GONE );
+      if ( mGooglePayButton   != null ) mGooglePayButton.setVisibility( View.GONE );
+      if ( mPayPalButton      != null ) mPayPalButton.setVisibility( View.GONE );
 
-      mCreditCardTextView.setText( R.string.kitesdk_payment_credit_card_button_text_free);
-      mCreditCardTextView.setOnClickListener( new View.OnClickListener()
+      mCreditCardButton.setText( R.string.kitesdk_payment_credit_card_button_text_free);
+      mCreditCardButton.setOnClickListener( new View.OnClickListener()
         {
         @Override
         public void onClick( View view )
@@ -251,14 +234,6 @@ public class DefaultPaymentFragment extends APaymentFragment
           submitOrderForPrinting( null, null, PaymentMethod.FREE );
           }
         } );
-      }
-    else
-      {
-      if ( mPayPalTextView != null ) mPayPalTextView.setVisibility( View.VISIBLE );
-      if ( mPayPalView     != null ) mPayPalView.setVisibility( View.VISIBLE );
-
-      mCreditCardTextView.setText( R.string.kitesdk_payment_credit_card_button_text);
-      mCreditCardTextView.setOnClickListener( this );
       }
     }
 
@@ -273,20 +248,13 @@ public class DefaultPaymentFragment extends APaymentFragment
     {
     // Both payment methods depend on us having the order price
 
-    if ( mOrderPricing != null )
-      {
-      if ( ( mPayPalTextView != null && view == mPayPalTextView ) ||
-           ( mPayPalView     != null && view == mPayPalView     ) )
-        {
-        onPayPalClicked( view );
-
-        return;
-        }
-      else if ( view == mCreditCardTextView )
-        {
-        onCreditCardClicked( view );
-
-        return;
+    if ( mOrderPricing != null ) {
+      if ( mGooglePayButton != null && view == mGooglePayButton ) {
+          onGooglePayClicked( view );
+        } else if ( mPayPalButton != null && view == mPayPalButton ) {
+          onPayPalClicked( view );
+        } else if (view == mCreditCardButton) {
+          onCreditCardClicked( view );
         }
       }
     }
@@ -339,8 +307,41 @@ public class DefaultPaymentFragment extends APaymentFragment
         }
 
       return;
-      }
 
+    } else if ( requestCode == ACTIVITY_REQUEST_CODE_GOOGLE_PAY ) {
+      switch ( resultCode ) {
+        case Activity.RESULT_OK:
+          PaymentData paymentData = PaymentData.getFromIntent( data );
+
+          if ( paymentData != null ) {
+            PaymentMethodToken token = paymentData.getPaymentMethodToken();
+
+            if ( token != null ) {
+              try {
+                final String stripeToken = new JSONObject( token.getToken() ).getString( "id" );
+                getPaymentActivity().submitOrderForPrinting( stripeToken, null, PaymentMethod.GOOGLE_PAY );
+              } catch ( JSONException e ) {
+                Log.i( LOG_TAG, "Error: " + e );
+              }
+            } else {
+              showErrorDialog( R.string.kitesdk_alert_dialog_message_no_payment_id );
+            }
+          } else {
+            showErrorDialog( R.string.kitesdk_alert_dialog_message_no_proof_of_payment );
+          }
+          break;
+
+        case Activity.RESULT_CANCELED:
+          break;
+
+        case AutoResolveHelper.RESULT_ERROR:
+          Status status = AutoResolveHelper.getStatusFromIntent( data );
+          if ( status != null ) Log.e( LOG_TAG, status.toString() );
+          break;
+        default:
+          // Do nothing.
+      }
+    }
 
     if ( mCreditCardAgent != null ) mCreditCardAgent.onActivityResult( requestCode, resultCode, data );
 
@@ -411,6 +412,63 @@ public class DefaultPaymentFragment extends APaymentFragment
 
   /*****************************************************
    *
+   * Starts the Google Pay payments client.
+   *
+   *****************************************************/
+  public void startGooglePayClient()
+  {
+    int walletConstant;
+
+    if (KiteSDK.getInstance(getActivity()).getEnvironmentName().equals("Live")){
+      walletConstant = WalletConstants.ENVIRONMENT_PRODUCTION;
+    } else {
+      walletConstant = WalletConstants.ENVIRONMENT_TEST;
+    }
+
+    mPaymentsClient = Wallet.getPaymentsClient(getActivity(), new Wallet.WalletOptions.Builder()
+      .setEnvironment(walletConstant)
+      .build());
+
+    // Checks if Google Pay is ready to pay
+
+    IsReadyToPayRequest request =
+      IsReadyToPayRequest.newBuilder()
+        .addAllowedPaymentMethod(WalletConstants.PAYMENT_METHOD_CARD)
+        .addAllowedPaymentMethod(WalletConstants.PAYMENT_METHOD_TOKENIZED_CARD)
+        .build();
+    Task<Boolean> task = mPaymentsClient.isReadyToPay(request);
+    task.addOnCompleteListener(
+      new OnCompleteListener<Boolean>() {
+        public void onComplete(Task<Boolean> task) {
+          try {
+            setGooglePayAvailable(task.getResult(ApiException.class));
+          } catch (ApiException exception) {
+            exception.getStatusCode();
+            setGooglePayAvailable(false);
+          }
+        }
+      });
+
+  }
+
+
+  /*****************************************************
+   *
+   * Sets availability of Google Pay.
+   *
+   *****************************************************/
+  private void setGooglePayAvailable(boolean available) {
+    if (available) {
+      mGooglePayButton.setOnClickListener( this );
+      mGooglePayButton.setVisibility(View.VISIBLE);
+    } else {
+      mGooglePayButton.setVisibility(View.GONE);
+    }
+  }
+
+
+  /*****************************************************
+   *
    * Called when pay by PayPal is clicked.
    *
    *****************************************************/
@@ -445,6 +503,24 @@ public class DefaultPaymentFragment extends APaymentFragment
       }
     }
 
+  /*****************************************************
+   *
+   * Called when pay by Google Pay is clicked.
+   *
+   *****************************************************/
+  public void onGooglePayClicked( View view ) {
+
+    final SingleCurrencyAmounts totalCost = getTotalCost();
+    if (totalCost != null) {
+      PaymentDataRequest request = createPaymentDataRequest(totalCost.getAmount().toString(),
+          totalCost.getCurrencyCode());
+      if (request != null) {
+        AutoResolveHelper.resolveTask( mPaymentsClient.loadPaymentData(request),
+            getPaymentActivity(), ACTIVITY_REQUEST_CODE_GOOGLE_PAY);
+      }
+    }
+  }
+
 
   /*****************************************************
    *
@@ -456,6 +532,48 @@ public class DefaultPaymentFragment extends APaymentFragment
     {
     getPaymentActivity().submitOrderForPrinting( authorisationProofOfPaymentFrom( paymentId ), accountId, paymentMethod );
     }
+
+
+  /*****************************************************
+   *
+   * Creates a PaymentDataRequest for Google Pay.
+   *
+   *****************************************************/
+
+  public PaymentDataRequest createPaymentDataRequest(String amount, String currencyCode) {
+    PaymentDataRequest.Builder request =
+        PaymentDataRequest.newBuilder()
+            .setTransactionInfo(
+                TransactionInfo.newBuilder()
+                    .setTotalPriceStatus(WalletConstants.TOTAL_PRICE_STATUS_FINAL)
+                    .setTotalPrice(amount)
+                    .setCurrencyCode(currencyCode)
+                    .build())
+            .addAllowedPaymentMethod(WalletConstants.PAYMENT_METHOD_CARD)
+            .addAllowedPaymentMethod(WalletConstants.PAYMENT_METHOD_TOKENIZED_CARD)
+            .setCardRequirements(
+                CardRequirements.newBuilder()
+                    .addAllowedCardNetworks(
+                        Arrays.asList(
+                            WalletConstants.CARD_NETWORK_AMEX,
+                            WalletConstants.CARD_NETWORK_DISCOVER,
+                            WalletConstants.CARD_NETWORK_VISA,
+                            WalletConstants.CARD_NETWORK_MASTERCARD))
+                    .build());
+
+    PaymentMethodTokenizationParameters params =
+        PaymentMethodTokenizationParameters.newBuilder()
+            .setPaymentMethodTokenizationType(
+                WalletConstants.PAYMENT_METHOD_TOKENIZATION_TYPE_PAYMENT_GATEWAY)
+            .addParameter("gateway", "stripe")
+            .addParameter("stripe:publishableKey", KiteSDK.getInstance(getActivity()).getStripePublicKey())
+            .addParameter("stripe:version", "5.0.0")
+            .build();
+
+    request.setPaymentMethodTokenizationParameters(params);
+
+    return request.build();
+  }
 
 
   ////////// Inner Class(es) //////////